+++ sources/packages/web/src/assets/css/student.scss
@@ -27,3 +27,39 @@
   color: $brand-blue;
   cursor: pointer;
 }
+
+.fa-app-header {
+  font-family: Noto Sans;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 20px;
+  line-height: 27px;
+  flex: none;
+  order: 0;
+  flex-grow: 0;
+  margin: 4px 0px;
+}
+.fa-app-header-1 {
+  @extend .fa-app-header;
+  color: #485363;
+  opacity: 0.5;
+}
+.fa-app-header-2 {
+  @extend .fa-app-header;
+  top: 31px;
+  letter-spacing: -0.2px;
+  color: #485363;
+}
+.fa-app-header-3 {
+  @extend .fa-app-header;
+  color: #2965c5;
+  line-height: 34px;
+}
+.img-background {
+  background-image: url("../images/icon_assistance.svg");
+  background-repeat: no-repeat;
+  background-size: contain;
+  width: 100%;
+  height: 100%;
+  min-height: 350px;
+}
