+++ sources/packages/web/src/composables/index.ts
@@ -11,4 +11,3 @@ export * from "./useAuthBCSC";
 export * from "./useStudentStore";
 export * from "./useFormatters";
 export * from "./useDesignationAgreement";
-export * from "./useStudentApplication";
