+++ sources/packages/web/src/views/aest/StudentApplicationView.vue
@@ -18,16 +18,10 @@
       }}
     </h2>
     <StudentApplication
-      :formName="selectedForm"
-      :data="initialData"
-      :readOnly="true"
-      @loaded="loadForm"
-    />
-    <StudentApplicationCommonActions
-      :isFirstPage="isFirstPage"
-      :isLastPage="isLastPage"
-      @wizardGoPrevious="wizardGoPrevious"
-      @wizardGoNext="wizardGoNext"
+      :selectedForm="selectedForm"
+      :initialData="initialData"
+      :programYearId="applicationDetail.applicationProgramYearID"
+      :isReadOnly="true"
     />
   </full-page-container>
   <router-view />
@@ -39,16 +33,13 @@ import { GetApplicationBaseDTO } from "@/types";
 import { ApplicationService } from "@/services/ApplicationService";
 import FullPageContainer from "@/components/layouts/FullPageContainer.vue";
 import HeaderNavigator from "@/components/generic/HeaderNavigator.vue";
-import { useStudentApplication } from "@/composables";
 import StudentApplication from "@/components/common/StudentApplication.vue";
-import StudentApplicationCommonActions from "@/components/common/StudentApplicationCommonActions.vue";
 
 export default {
   components: {
     FullPageContainer,
     HeaderNavigator,
     StudentApplication,
-    StudentApplicationCommonActions,
   },
   props: {
     studentId: {
@@ -64,14 +55,6 @@ export default {
     const applicationDetail = ref({} as GetApplicationBaseDTO);
     const initialData = ref({});
     const selectedForm = ref();
-    const isFirstPage = ref(true);
-    const isLastPage = ref(false);
-    // TODO:ANN- CHECK HERE THE PROGRAMYEARID, 2ND parametre of useStudentApplication
-    const {
-      formLoaded,
-      wizardGoPrevious,
-      wizardGoNext,
-    } = useStudentApplication(true, 1);
 
     onMounted(async () => {
       applicationDetail.value = await ApplicationService.shared.getApplicationDetail(
@@ -82,19 +65,10 @@ export default {
       initialData.value = applicationDetail.value.data;
     });
 
-    const loadForm = (form: any) => {
-      formLoaded(form);
-    };
-
     return {
       applicationDetail,
       initialData,
       selectedForm,
-      isFirstPage,
-      isLastPage,
-      loadForm,
-      wizardGoPrevious,
-      wizardGoNext,
       AESTRoutesConst,
     };
   },
